+++ src/main/java/net/frozenblock/freezeframe/component/tooltip/client/ClientFilmTooltip.java
@@ -144,7 +144,7 @@ public class ClientFilmTooltip implements ClientTooltipComponent {
 	}
 
 	private static int getContentXOffset(int tooltipWidth) {
-		return (tooltipWidth - GRID_WIDTH) / 2;
+		return (tooltipWidth) / 2 - (Math.max(tooltipWidth, GRID_WIDTH) / 2);
 	}
 
 	@Override
@@ -190,10 +190,6 @@ public class ClientFilmTooltip implements ClientTooltipComponent {
 		final int left = x + getContentXOffset(w);
 		final int photoDrawX = left + ((w / 2) - (PHOTOGRAPH_RENDER_SIZE / 2));
 		PhotographRenderer.blit(photoDrawX, y, 0, 0, graphics, this.photographId, PHOTOGRAPH_RENDER_SIZE, PhotographRenderer.FrameType.FILM_EMBED);
-
-		// TODO: extract arrows
-		if (this.hasMultiplePhotographs) {
-		}
 
 		this.extractPhotographTooltips(left, y + TOOLTIP_HEIGHT, font, graphics);
 	}