+++ CHANGELOG.md
@@ -9,5 +9,6 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 - Added separate textures for the Camera item, resolving an issue that locked mipmapping to 3 levels.
 - Resetting a scope item's zoom to default now plays multiple increment sounds in succession instead of one, relative to the amount of zoom steps away from the default.
 - Fixed an issue that would cause an error about the `ContainerComponentManipulators` class to log, despite working as intended.
+- Fixed the Film tooltip to now be centered regardless of the tooltip's width.
 - The Film Capacity Upgrade Recipe now supports a custom Source, Material, and Output item like other Vanilla transmute recipes.
 - The Film Capacity Upgrade Recipe now shows up in the Crafting Recipe Book.